Fix handling of special Windows path prefixes
Roughly speaking, when parsing a value of type path on Windows, any path
that starts with \\ had an additional \ appended to it. For example:

    def to_path [p: path] { $p }
    to_path \\.\CON
    # => \\.\CON\
    to_path \\server\share
    # => \\server\share\

More precisely, these include UNC paths (\\server\share), device paths
(\\.\device) and verbatim paths (starting with \\?\). In these cases,
std::path::Path::components() parses the whole path as a prefix, but
still appends a std::path:component::RootDir, which caused the trailing
backslash to be introduced when re-assembling the comonents into a
PathBuf.